BUG: Fix nodes not getting removed in vtkMRMLSequenceNode::SetDataNodeAtValue
When SetDataNodeAtValue is called, the new data node is deep-copied into the internal sequence scene, and replaces the old node in the index list if one exists. The old node however is not removed from the scene, causing the number of nodes to grow larger every time SetDataNodeAtValue was called for a value with an existing data node.

Fixed by removing the old node from the scene if one already exists for the specified value.
This commit also changes the node pointer in IndexEntryType to a vtkWeakPointer.

Fixes #7328