ENH: Update codespell workflow to use latest version of codespell
To follow-up on 866fe39a0 (DOC: Fix typos in various subdirectories (#6483)),
which temporarily addressed the issue of an outdated codespell version in
our workflow, this commit revises our approach.

Previously, we had worked around the problem by adopting a forked action,
"Slicer/actions-codespell" which included an updated requirements file.
However, this approach proved to be suboptimal because it lacked systematic
version updates for codespell.

In this commit, we revert to using the official "actions-codespell" action,
which now leverages "codespell[toml]>=2.2.4," ensuring the consistent use
of an up-to-date version. This change guarantees that our workflow will
always benefit from the latest improvements and fixes provided by codespell.