ENH: add support for asynchronous dicom scalar volume load
Before, the `DICOMScalarVolumePlugin` only supported working
with file system dicom data (so called part10 binary files, typicaly
with the .dcm extension).  These files are read synchronously and the
results are stored in the volume node when reading from disk is complete.
This approach still exists and is the default.

With this change and corresposnding changes to `ctkDICOMDatabase`
and associated changes in this branch:
https://github.com/pieper/CTK/tree/virtualize-database
it is now possible for external code to add the ability to populate
the image data of a `vtkMTMLScalarVolumeNode` asynchronously, for example
as a result of network downloads.

The new feature allows the MRML node to be created during the `load`
method, and then incrementally updated as downloads complete.