BUG: Fix language translation of subject hierarchy plugins
Translation context is the Python package name (parent folder name that contains __init__.py).
Subject hierarchy abstract plugin was moved to a different folder during installation, therefore the context that was determined in the source tree did not work in the install tree. It is fixed by using the same folder structure in the source and install tree.
Where modules specified their plugins, __init__.py was missing and therefore the context name was not correct (it was the parent folder of SubjectHierarchyPlugins). It is fixed by adding the missing __init__.py file.

see https://github.com/Slicer/SlicerLanguagePacks/issues/21

fixes https://github.com/Slicer/SlicerLanguagePacks/issues/52

Co-authored-by: Andras Lasso <lasso@queensu.ca>