ENH: dicom url retries and code cleanup
Catch if the returned data is not as expected and if
so do a retry.  Based on testing with a Google DICOMweb
store about 5-10 request were incomplete on a several hundred
slice dataset but no error code was reported, just truncated data.
There's no clear reason to why these failures occur, but
retrying implemented here has always resulted in good data
during testing so far.