STYLE: Update Python docstrings adding newline after last paragraph
Fixes introduced in this commit were applied using the Ruff
configuration updated in the next commit.

PEP 257 recommends that the closing quotes of a multi-line docstring
be on their own line, for consistency and compatibility with documentation
tools that may need to parse the docstring.

See https://docs.astral.sh/ruff/rules/new-line-after-last-paragraph/