BUG: Resolve del method override issue in AutoComplete segment editor effect
This commit rectifies the linter error concerning the __del__ method in the
AutoComplete segment editor effect. The specific issue was that the __del__
method expects one parameter, but two were given. To resolve this, the call
to the superclass __del__ has been removed, as it is not implemented.

Note that while this correction resolves the immediate warning, there may still
be issues with proper garbage collection for some effects. A proposal to address
the root cause is documented in Slicer#7392

It also updates the Ruff configuration, re-enabling the PLE0302 rule to
ensure that __del__ functions are called with the expected number of
parameters.