ENH: Simplify Endoscopy math helper functions API to only return value
Supporting an API allowing to get output using both "return by value"
and "pass mutable object as argument" is error-prone.

Considering the returned value is only few bytes (e.g 32 bytes for wxyz
orientation and 128 bytes for a 4x4 matrix), the added complexity it not
worth it.

To adhere to the principle of least surprise and make the code more
readable, this commit simplifies the API and updates the code accordingly.