BUG: Fix rebuild of Endoscopy flythrough when modifying the input curve
This commit addresses an issue introduced during the integration of the pull request
associated with commit 717fd6f2b (ENH: Add support for managing orientation keyframes
in Endoscopy).

The logic that was inadvertently removed is now reinstated to prevent unintended modification
events triggered by changes to the input curve. The variable "ignoreInputCurveModified"
is reintroduced to manage and prevent a second modification event.

Additionally, this commit addresses typos in comments.

Co-authored-by: Jean-Christophe Fillion-Robin <jchris.fillionr@kitware.com>