BUG: Recompute EndoscopyLogic fields immediately instead of lazily
The logic for when to loadCameraOrientations and recompute EndoscopyLogic
fields was not working in all cases.  The lazy approach to loading and
recomputing could have been modified to fix this, but we did not pursue this
approach because the lazy approach has a flaw: the user interface (e.g. number
of frames available) should be available immediately, not lazily (e.g. only
once the user starts a flythrough).

(cherry picked from commit 5e45824d5cc654a8277205921fb4a1ed129167bb)