ENH: Simplify EndoscopyLogic API to set and get input curve CameraOrientations
This allows to set and get the input curve CameraOrientations independently
of the EndoscopyWidget.

To support this it also updates the `interpolateOrientations` to require
a cameraOrientations parameter. This allows to remove the need for maintaining
a dedicated "cameraOrientations" instance variable.

It also renames the functions:
* `loadCameraOrientations` to `getCameraOrientationsFromInputCurve`
* `saveCameraOrientations` to `setInputCurveCameraOrientations`

(cherry picked from commit b6975a4f5a3703e8b75ae447519e12e71204e1de)