BUG: Fix loading of sequence nodes
When a sequence was saved into a scene and then that scene was loaded then browsing did not work.
The problem was caused by a typo that set missingItemMode to invalid and thereby disabling copying of sequence content into the proxy node.

Fixed the typo, made setting of missingItemMode more robust (if invalid string is found then the mode is not changed just a warning is logged), and added automated tests to capture similar errors in the future.

fixes #7440

(cherry picked from commit 140d58aa18a4c293634a5cbfafd121254fc9e7f9)