BUG: Update VTK to to fix vtkSMPToolsAPI static initialization order issue
This addresses a crash related to the static initialization order 'fiasco'
in vtkSMPToolsAPI. It ensures proper deletion of the vtkSMPToolsAPI singleton
once the last translation unit referencing it has been unloaded.

It also ensures backend-specific statics are properly initialized.
when the first translation unit is loaded.

This update ensures that the SMP backend is cleaned up at the appropriate
time, fixing the following macOS tests that started to fail following commit
28dca75dc (ENH: Enable TBB as the default VTK SMP implementation on all platforms):
* `py_nomainwindow_SegmentationWidgetsTest1`
* `py_sceneImport2428`
* `py_SegmentStatistics`
* `py_UtilTest`

List of changes:

$ git shortlog 4bfb0f049a..46201478cd --no-merges
Jean-Christophe Fillion-Robin (3):
      [Backport MR-10751] BUG: Resolve crash by fixing vtkSMPToolsAPI static initialization order issue
      [Backport MR-10751] ENH: Simplify vtkSMPToolsAPIInitialize managing pointer and counter locally
      [Backport MR-10751] BUG: Resolve crash by fixing initialization of backend-specific statics