ENH: Load scalar volumes from DICOMweb
Previous work allowed populating the ctkDICOMDatabase from
dicomweb and retrieving frames.  However the tag cache performance
was a bottleneck for when there are several hundred studies
in the database.

This change allows the scalar volume plugin to get
data from a a proxy database, which can pull the metadata
on demand from a dicom store.

This change requires proxy code that is currently available
in the DIOCMLogic package.