BUG: Fix adding of new sequence item when going to previous item
When a sequence contained only a single item and clicked the previous frame button in the sequence browser, a new item was added to the sequence.

The problem was with incorrect computation when number of items was the same as the negative offset (e.g., offset by -1 frame when number of frames was 1).