BUG: Fix support for registering parameterNodeWrapper plugins
Revisit the support for module-specific parameterNodeWrapper introduced
in 19f787127 (BUG: Register parameterNodeWrapper plugins in corresponding
module). The goal is to prevent the installation of existing plugins from
being overwritten by those provided by other modules.

To address this, the commit organizes each module-specific parameterNodeWrapper
in a package following the convention "<moduleName>ParameterNodeWrapperPlugins".

This approach leverages Python's built-in facilities [1] for looking up packages
that match a specific convention. It ensures that the plugins will be registered
consistently, whether they originate from the Slicer core or a Slicer extension.

[1] https://packaging.python.org/en/latest/guides/creating-and-discovering-plugins/#using-naming-convention

Co-authored-by: Andras Lasso <lasso@queensu.ca>