COMP: Fix generation of additional launcher settings for list-based settings
This commit addresses issues identified after e9f28d6c8 (`ENH: Fixed
AdditionalLauncherSettings.ini for extensions with dependencies`) and
commit b1efab8a6 (`COMP: Expand <CMAKE_CFG_INTDIR> in extension additional
launcher settings`) the use of the `string(REPLACE ...)` command was
introduced.

However, using `string(REPLACE ...)` with a variable set as a list results
in a concatenated string with the replaced values. This commit addresses
this by transitioning to the `TRANSFORM` action, capable of replacing values
individually within a list.

Lastly, this commit addresses the missing replacement of `<CMAKE_CFG_INTDIR>`
associated with the `PYTHONPATH_LAUNCHER_BUILD` settings.