COMP: Fix dependent extension paths used in additional launcher settings
This commit fixes the regular expression used for replacing `$(Configuration)`
with `${CMAKE_CFG_INTDIR}` so that the placeholder variable is ultimately
replaced through the execution of the custom command associated with the
`ConfigureAdditionalLauncherSettings` custom target.

It follows up on e9f28d6c8 (`ENH: Fixed AdditionalLauncherSettings.ini
for extensions with dependencies`) were the original regular expression
was introduced.