BUG: Update VTK backporting OpenVR rendering fix when HMD inside the volume
This addresses proper rendering, particularly by `vtkOpenGLGPUVolumeRaycastMapper`,
when the head-mounted display (HMD) is positioned inside the volume. The issue is
resolved by ensuring correct computation of the plane frustum when the camera is
tracking the HMD and is positioned inside the volume.

The changes listed below impact the following VTK modules:
* `Rendering/VR`
* `Rendering/OpenVR`
* `Rendering/OpenXR`

List of VTK changes:

```
$ git shortlog c69cf77a6..7767cabe4 --no-merges
Sankhesh Jhaveri (3):
      [Backport MR-10783] VR: Move the TrackHMD flag to the camera
      [Backport MR-10783] VR: Deprecate vtkVRRenderWindow::TrackHMD
      [Backport MR-10783] VR: Provide stereo projection transforms for rendering
```

Co-authored-by: Sankhesh Jhaveri <sankhesh.jhaveri@kitware.com>
Co-authored-by: Lucas Gandel <lucas.gandel@kitware.com>