ENH: Introduce Visual DICOM Browser with CTK Update
This enhancement includes an update to CTK and introduces the `ctkDICOMVisualBrowserWidget`
to augment DICOM exploration, query, and retrieval capabilities within the DICOM
module. The widget is seamlessly integrated into the Slicer DICOM module and is
labeled as experimental.

For a comprehensive overview and future plans, refer to the roadmap at
https://github.com/commontk/CTK/issues/1162.

By default, the widget is disabled, and users can access the familiar `ctkDICOMBrowser`
when opening the `DICOM` module. To enable the experimental feature, users can
toggle the option in the dropdown menu of the `Show DICOM database` pushbutton
in the `DICOM` module UI.

In `SlicerDICOMBrowser`, this introduces an instance of `ctkDICOMVisualBrowserWidget`
alongside the existing `ctkDICOMBrowser` instance. When users activate the
experimental feature, widget visibilities are adjusted accordingly. Both widgets
share the same DICOM folder set in the DICOM settings.

List of CTK changes:

```
$ git shortlog 45f33c81..88ff72b9 --group=author --group=trailer:co-authored-by --no-merges
Andras Lasso (1):
      ENH: Add Visual DICOM Browser (#1165)

Davide Punzo (1):
      ENH: Add Visual DICOM Browser (#1165)

Jean-Christophe Fillion-Robin (1):
      ENH: Add Visual DICOM Browser (#1165)
```

Co-authored-by: Andras Lasso <lasso@queensu.ca>
Co-authored-by: Jean-Christophe Fillion-Robin <jchris.fillionr@kitware.com>