ENH: Allow dragging of splashwindow
Splashwindow is a topmost window that is shown during application startup. Since it is a topmost window that appears in the screen center above all other applications, it may make the computer hard to use for anything else.

This commit makes the splashwindow less obstructive: The splashwindow can be moved by clicking on it (anywhere) and dragging it out of the way. Clicking it also removes "topmost" attribute from splashwindow, so it is no longer forced to appear above everything else.