ENH: Support dragging of the splashwindow just after clicking on it
Follow-up of edfbdb492 (ENH: Allow dragging of splashwindow), this commit
further improve the `DraggableWidgetEventFilter` to support dragging the
Splashwindow without having to release the mouse button and re-pressing it.

Since just after the mouse button is pressed and before it is released, mouse
events are associated with a "QWidgetWindow" instance named after "<objectname>Window"
if the widget object name is set or after "<classname>ClassWindow" if the object name is empty
(See `QWidgetWindow::updateObjectName`).

To ensure events from either are being filtered, the event filter should be installed at the
application level.