BUG: Fix meta-object system registration of qSlicerIO::fileType return type
To ensure correct lookup of the return type of the `fileType` function registered
with the meta-object system (MOC), this commit updates the function signature.
Previously, it was registered as "IOFileType," and this caused issues with
PythonQt, which expected "qSlicerIO::IOFileType" for proper identification.

This change eliminates the need to separately register the unqualified and
potentially conflicting shorter version:

```cpp
qRegisterMetaType<qSlicerIO::IOFileType>("IOFileType");
```

It also fixes the following error:

```
reader = slicer.qSlicerSceneBundleReader()
reader.fileType()
Traceback (most recent call last):
  File "<console>", line 1, in <module>
ValueError: Called fileType() -> IOFileType, return type 'IOFileType' is ignored because it is unknown to PythonQt. Probably you should register it using qRegisterMetaType() or add a default constructor decorator to the class.
```