ENH: SImplify DraggableWidgetEventFilter
Instead of attaching an event filter on the application and then analyzing the source of the events, it is cleaner to attach the event filter to the popup.
The issue of changing the window flags while processing the button click events can be handled more cleanly and robustly by changing the window flag in a timer callback.