BUG: Ensure packaging of the latest d3dcompiler_47.dll for SlicerVirtualReality
This commit focuses on updating the Qt install rules to ensure the packaging of the
latest version of "d3dcompiler_47.dll". The proper functioning of
"Microsoft.Holographic.Remoting.OpenXr" within the SlicerVirtualReality extension
with OpenXR remoting enabled relies on this update for successful scene rendering.

By setting the `WindowsSdkDir` environment variable, the "windeployqt" tool is
instructed to look up and deploy the library available on the system. This action
prevents the deployment of an older version that may be installed alongside Qt 5.15.2
libraries when Qt is installed using the installer.

Co-authored-by: Alicia Pose <alicia10pose@gmail.com>
Co-authored-by: Amaia Iribar <airibar@vicomtech.org>
Co-authored-by: Felix von Haxthausen <37213426+felixvh@users.noreply.github.com>
Co-authored-by: Monica Garcia-Sevilla <90038097+monicagsevilla@users.noreply.github.com>