BUG: Ensure launcher scripts installed from wheels are accessible
This commit addresses an issue where launcher scripts installed from wheels
were not correctly added to the PATH. Specifically, it updates the PATH for
both "Slicer" and "PythonSlicer" launcher settings in the build and install
tree. This adjustment ensures that convenience executables created after
installing wheels with specified entry-points such as "project.scripts" or
"project.gui-scripts" can be easily located and accessed within the project
environment.