BUG: Fix loading of empty segmentation
NRRD file cannot be created without voxel data, so an 1x1x1 array was used as image data.
However, when the empty segmentation was loaded then this single-voxel geometry was treated as a valid geometry.
Fixed by detecting this special case of a single voxel in the segmentation, and considering the geometry invalid in this case.

fixes #7576