ENH: Make translations in Python scripts more robust
If a Python string with placeholders (e.g., Loading {filename}) is translated incorrectly so that the placeholder is not in the translated string (e.g., Carga {nombre de archivo}) then currently an exception is raised and execution of the Python code stops.

Fixed by only using the translated string if the placeholders match. Otherwise a warning is logged and the translation is not used.

fixes #7235