BUG: Fix subject hierarchy consolidation after loading a scene
When loading a scene, subject hierarchy items were not consolidated automatically.
This did not cause issues when using the main Slicer scene, as the subject hierarchy plugin logic did it;
but it is still an issue when processing an independently created scene object.

The solution is to call ResolveSubjectHierarchy method at the end of Connect method.