ENH: Switch extension index entry format from "s4ext" to "json"
This commit implements changes to the Extensions build-system,
transitioning from `.s4ext` files to `.json` files for extension
index entries. Additionally, it updates the ExtensionWizard and
module templates to reflect this change and supports contributing
`.json` files to the ExtensionsIndex.

As the category is now expected to be defined in the extension catalog
entry file, the support for setting the "Category" directly from
the ExtensionWizard UI has been removed.

While a .s4ext file is still generated in the built-tree and included
in the extension package, this is considered an implementation detail
to be addressed in subsequent commits.

The motivations behind these changes include:
* Eliminate redundant and unused information from the "description" file.
* Simplify programmatic parsing of the "description" files
* Decouple the metadata organized in the extension `CMakeLists.txt` from
  the ones organized in this repository and used to drive the build of extensions.
* Enable Slicer maintainers to define and update the extension "category"
  and "enabled" metadata independently of the upstream extension sources.