ENH: De-duplicate setting of extension metadata
Anticipating the simplification of the metadata used to describe an extension
Simplifies handling of metadata
catalog entry, this removes the parsing and handling of metadata already defined
in the extension CMakeLists.txt by the extension index build-system.

The following metadata values are now extracted from the `<extensionname>.s4ext`
file locally generated in the extension build tree:
- "category"
- "contributors"
- "enabled"
- "depends" (runtime dependencies)
- "description"
- "iconurl"
- "homepage"
- "screenshots"

The only metadata values propagated from the Slicer/ExtensionsIndex files
are the following:
- "scm", "scmurl" and "scmrevision"
- "depends" (built-time dependencies)
- "build_subdirectory"

The extension name is still derived from the catalog entry filename.