ENH: Add DICOM patcher rule to split ultrasound series
On some ultrasound systems, all ultrasound images in a study get saved with the same series instance UID.
This makes it very difficult to browser images, because when the series is opened then all the (potentially dozens or even hundreds) acquisitions are loaded, each one as a new sequence.

The new rule allows the user to split an ultrasound series based on the "Instance number" attribute. Each instance number within a series is split out into a new series.