Show warning window if data loss occurs when saving to tiff (#7655)
* BUG: Fix creation of translation template file for single language

When a language was specified for updating translations and no translation file was found for that language
then a new translation file was generated for en-US language. This commit changes this to generate translation file
for the specified language.

* ENH: Add helper function for MRML translation with placeholders

* ENH: Display warning when saving into TIFF format

Adds warning for TIFF file format saving. Prevents potential data loss.
Made translatable all messages that may be relevant to users.

see #7588

Co-authored-by: Andras Lasso <lasso@queensu.ca>

---------

Co-authored-by: Aram Harutyunyan <aram.harutyunyan@softhenge.com>