ENH: Allow aliases for NRRD "type" header in postNRRD
The NRRD definition lists various aliases
to specify the for the "short" datatype in
the "type" nrrd header field (see
https://teem.sourceforge.net/nrrd/format.html#type).

Pynnrd as a widely used nnrd library
e.g. creates NRRD headers with short
datatypes identified as "int16" and
does not allow to override this.
Consequently, a POST request to the
slicer/volume endpoint will always fail
with data from pynnrrd.

This commit relaxes the type check to
also allow aliases of "short" to pass.