ENH: Improve terminology selector
Allow deselecting anatomic region and anatomic region modifier ("None" and "No region modifier" items are added to the comboboxes). Before this commit, once a region or modifier was selected, it was not possible to deselect it.

Add the anatomic region to the default generated name, because often it is the only difference between segment names (e.g., "bleeding in Cerebellum" and "bleeding in Subdural space").

To reduce occurrences of undefined (gray) color: if color is not specified for the selected type modifier then use the color of the non-modified type.

Added a few missing translation functions to allow full internationalization.