ENH: Allow disabling histogram (2D views) interactions of ThresholdEffect
This is useful in custom workflows where threshold preview is needed but interactive adjustment of threshold range must be prevented.

It is not added as a effect parameter because then the setting would have been always saved in the scene and exposed on the GUI, adding complexity to the scene and the GUI - for supporting a very rare use case.