ENH: Improve terminology selector speed
When the type selector was populated, all the type IDs were retrieved first, and then the type objects were looked up based on these IDs.
This lookup is very slow if there are tens of thousands of types in the terminology.

Improved performance by returning not just all the type IDs but also all the corresponding type objects, therefore avoiding the costly lookup for each type.