ENH: Make it optional to force right-handed IJK coordinate system
For troubleshooting image reading, it is useful if it is possible to read an image with the exact coordinate system axis directions stored in the file. Therefore, added ForceRightHandedIJKCoordinateSystem flag to vtkMRMLVolumeArchetypeStorageNode (set to true by default) to allow developers to customize the behavior.

Moved the private utility functions to the vtkMRMLVolumeNode, as they can be useful for manual normalization of coordinate system axes directions.

Improved the vtkMRMLVolumeArchetypeStorageNodeTest1 to test that the exact same image content is displayed at the same physical location.