BUG: Fix warnings when vtkMRMLModelNode has no mesh data
Fixes warnings during rendering such as:

> Algorithm vtkPassThrough (00000223357FFBC0) returned failure for request: vtkInformation (000002233CEA6520)

Fixed by enabling AllowNullInput on the vtkPassThrough in vtkMRMLModelDisplayNode.

This commit also fixes an issue where the views would not be immediately rendered when vtkMRMLModelNode::SetAndObserveMesh(nullptr) was called for a model node that previously had a mesh.
Fixed by not returning early and accounting for IsModelDisplayable in OnMRMLDisplayableModelNodeModifiedEvent.