BUG: Fix tensor image reading
Reading of tensor images (such as DTIBrain sample data set) crashed Slicer.
The root cause was that the image used a left-handed coordinate system, the recently added image orientation normalization function tried to flip the image, but the underlying vtkImageReslice filter crashes if the scalar point data in the input image is nullptr (which is the case for tensor images, where the tensor point data is valid but scalar point data is invalid).

Fixed by setting the array that contains the data to the scalar array for the time of running vtkImageReslice.

Fixes py_CompareVolumes test failure.